edit and save multiple availabilityslots at the same time saves to database, still some issues when fetching timeslots with time displayed.
